requirements.txt
requests
python-dotenv
oauthlib
requests_oauthlib